+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [API Documentation on Read the Docs](https://echovr-api.readthedocs.io/en/latest/index.html)
 
 ## [0.1.1] - 2018-11-06
 ### Fixed