+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0] - 2018-11-18
 ### Added
 - [API Documentation on Read the Docs](https://echovr-api.readthedocs.io/en/latest/index.html)
 - Support for new API response properties in Echo VR Combat release patch
@@ -33,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Simple system for fetching JSON from the API
 - Simple object model exposing the basic functionality of the API
 
-[Unreleased]: https://github.com/ajedi32/echovr-api/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/ajedi32/echovr-api/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/ajedi32/echovr-api/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/ajedi32/echovr-api/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/ajedi32/echovr-api/compare/v0.0.1...v0.1.0